@@ -8,30 +8,54 @@
 #include <llvm/IR/InstrTypes.h>
 #include "generator.h"
 
+Generator::Generator() {
+    module = std::make_unique<Module>("trnp2", context);
+    builder = std::make_unique<IRBuilder<>>(context);
+
+    passmgr = std::make_unique<legacy::FunctionPassManager>(module.get());
+    passmgr->add(createInstructionSimplifierPass());
+    passmgr->add(createInstructionCombiningPass());
+    passmgr->add(createReassociatePass());
+    passmgr->add(createDeadCodeEliminationPass());
+    passmgr->add(createDeadInstEliminationPass());
+    passmgr->add(createDeadStoreEliminationPass());
+    passmgr->add(createConstantHoistingPass());
+    passmgr->add(createLoadCombinePass());
+    passmgr->add(createAggressiveDCEPass());
+    passmgr->add(createLoopLoadEliminationPass());
+    passmgr->add(createConstantPropagationPass());
+    passmgr->add(createGVNHoistPass());
+    passmgr->add(createCFGSimplificationPass());
+    passmgr->add(createLoopSimplifyCFGPass());
+    passmgr->add(createMemCpyOptPass());
+
+    passmgr->doInitialization();
+}
+
 void Generator::use_io() {
     io_using = true;
 
-    table.insert(std::pair<std::string, Value *>("int_out_format", builder->CreateGlobalStringPtr("%i\n", "int_out_format")));
-    table.insert(std::pair<std::string, Value *>("float_out_format", builder->CreateGlobalStringPtr("%f\n", "float_out_format")));
-    table.insert(std::pair<std::string, Value *>("str_out_format", builder->CreateGlobalStringPtr("%s\n", "str_out_format")));
+    table.emplace("int_out_format", builder->CreateGlobalStringPtr("%i\n", "int_out_format"));
+    table.emplace("float_out_format", builder->CreateGlobalStringPtr("%f\n", "float_out_format"));
+    table.emplace("str_out_format", builder->CreateGlobalStringPtr("%s\n", "str_out_format"));
 
-    printfArgs.push_back(Type::getInt8PtrTy(context)); // create the prototype of printf function
+    printfArgs.emplace_back(Type::getInt8PtrTy(context)); // create the prototype of printf function
     printfType = FunctionType::get(Type::getInt32Ty(context), printfArgs, true);
     printf = module->getOrInsertFunction("printf", printfType);
 
-    table.insert(std::pair<std::string, Value *>("float_in_format", builder->CreateGlobalStringPtr("%d", "float_in_format")));
+    table.emplace("float_in_format", builder->CreateGlobalStringPtr("%d", "float_in_format"));
 
-    scanfArgs.push_back(Type::getInt8PtrTy(context)); // create the prototype of scanf function
+    scanfArgs.emplace_back(Type::getInt8PtrTy(context)); // create the prototype of scanf function
     scanfType = FunctionType::get(Type::getInt32Ty(context), scanfArgs, true);
     scanf = module->getOrInsertFunction("scanf", scanfType);
 }
 
-void Generator::generate(Node *n) {
+void Generator::generate(const std::shared_ptr<Node>& n) {
     switch(n->kind) {
         case Node::NEW: {
-            if (n->o1 != NULL) { // new array
+            if (n->o1 != nullptr) { // new array
                 if (n->o1->kind == Node::ARRAY) {
-                    Node *arr = n->o1;
+                    std::shared_ptr<Node> arr = n->o1;
                     generate(arr->o1);
                     Value *elements_count_val = stack.top();
                     stack.pop();
@@ -39,37 +63,49 @@ void Generator::generate(Node *n) {
                     // get number of elements in array
                     int64_t elements_count = 0;
                     if (llvm::ConstantInt *CI = dyn_cast<llvm::ConstantInt>(elements_count_val)) {
-                        if (CI->getBitWidth() <= 32)
+                        if (CI->getBitWidth() <= 32) {
                             elements_count = CI->getSExtValue();
+                        }
                     }
 
                     switch (arr->value_type) {
-                        case Node::integer: // integer array
-                            table.insert(
-                                    std::pair<std::string, Value *>(
-                                            n->var_name,
-                                            builder->CreateAlloca(
-                                                    ArrayType::get(
-                                                            Type::getInt32Ty(context),
-                                                            static_cast<uint64_t>(elements_count)
-                                                    ),
-                                                    nullptr,
-                                                    n->var_name + "_ptr")
+                        last_vars.emplace_back(n->var_name);
+                        case Node::INTEGER: // integer array
+                            table.emplace(
+                                    n->var_name,
+                                    builder->CreateAlloca(
+                                            ArrayType::get(
+                                                    Type::getInt32Ty(context),
+                                                    static_cast<uint64_t>(elements_count)
+                                            ),
+                                            nullptr,
+                                            n->var_name + "_ptr"
                                     )
                             );
                             break;
-                        case Node::floating: // float array
-                            table.insert(
-                                    std::pair<std::string, Value *>(
-                                            n->var_name,
-                                            builder->CreateAlloca(
-                                                    ArrayType::get(
-                                                            Type::getDoubleTy(context),
-                                                            static_cast<uint64_t>(elements_count)
-                                                    ),
-                                                    nullptr,
-                                                    n->var_name + "_ptr"
-                                            )
+                        case Node::FLOATING: // float array
+                            table.emplace(
+                                    n->var_name,
+                                    builder->CreateAlloca(
+                                            ArrayType::get(
+                                                    Type::getDoubleTy(context),
+                                                    static_cast<uint64_t>(elements_count)
+                                            ),
+                                            nullptr,
+                                            n->var_name + "_ptr"
+                                    )
+                            );
+                            break;
+                        case Node::USER: // float array
+                            table.emplace(
+                                    n->var_name,
+                                    builder->CreateAlloca(
+                                            ArrayType::get(
+                                                    PointerType::get(user_types.at(n->user_type).first, 0),
+                                                    static_cast<uint64_t>(elements_count)
+                                            ),
+                                            nullptr,
+                                            n->var_name + "_ptr"
                                     )
                             );
                             break;
@@ -77,26 +113,35 @@ void Generator::generate(Node *n) {
                 }
             }
             else { // just variable
+                last_vars.emplace_back(n->var_name);
                 switch (n->value_type) {
-                    case Node::integer: // int variable
-                        table.insert(
-                                std::pair<std::string, Value *>(
-                                        n->var_name,
-                                        builder->CreateAlloca(
-                                                Type::getInt32Ty(context),
-                                                nullptr,
-                                                n->var_name + "_ptr")
+                    case Node::INTEGER: // int variable
+                        table.emplace(
+                                n->var_name,
+                                builder->CreateAlloca(
+                                        Type::getInt32Ty(context),
+                                        nullptr,
+                                        n->var_name + "_ptr"
                                 )
                         );
                         break;
-                    case Node::floating: // float variable
-                        table.insert(
-                                std::pair<std::string, Value *>(
-                                        n->var_name,
-                                        builder->CreateAlloca(
-                                                Type::getDoubleTy(context),
-                                                nullptr,
-                                                n->var_name + "_ptr")
+                    case Node::FLOATING: // float variable
+                        table.emplace(
+                                n->var_name,
+                                builder->CreateAlloca(
+                                        Type::getDoubleTy(context),
+                                        nullptr,
+                                        n->var_name + "_ptr"
+                                )
+                        );
+                        break;
+                    case Node::USER: // user-type variable
+                        table.emplace(
+                                n->var_name,
+                                builder->CreateAlloca(
+                                        user_types.at(n->user_type).first,
+                                        nullptr,
+                                        n->var_name + "_ptr"
                                 )
                         );
                         break;
@@ -105,83 +150,379 @@ void Generator::generate(Node *n) {
             break;
         }
         case Node::INIT: { // create and initialize the variable
+            last_vars.emplace_back(n->var_name);
             switch (n->value_type) {
-                case Node::integer: // integer variable
-                    table.insert(
-                            std::pair<std::string, Value *>(n->var_name,
-                                                            builder->CreateAlloca(
-                                                                    Type::getInt32Ty(context),
-                                                                    nullptr,
-                                                                    n->var_name + "_ptr")
+                case Node::INTEGER: // integer variable
+                    table.emplace(
+                            n->var_name,
+                            builder->CreateAlloca(
+                                    Type::getInt32Ty(context),
+                                    nullptr,
+                                    n->var_name + "_ptr"
                             )
                     );
                     break;
-                case Node::floating: // float variable
-                    table.insert(
-                            std::pair<std::string, Value *>(n->var_name,                                                            builder->CreateAlloca(
-                                                                    Type::getDoubleTy(context),
-                                                                    nullptr,
-                                                                    n->var_name + "_ptr")
+                case Node::FLOATING: // float variable
+                    table.emplace(
+                            n->var_name,
+                            builder->CreateAlloca(
+                                    Type::getDoubleTy(context),
+                                    nullptr,
+                                    n->var_name + "_ptr"
+                            )
+                    );
+                    break;
+                case Node::USER: // user-type variable
+                    table.emplace(
+                            n->var_name,
+                            builder->CreateAlloca(
+                                    user_types.at(n->user_type).first,
+                                    nullptr,
+                                    n->var_name + "_ptr"
                             )
                     );
                     break;
             }
 
-            generate(n->o1); // generate initial value
-            Value *val = stack.top(); // take it from the stack
-            stack.pop(); // erase it from the stack
+            if (n->o1->value_type == Node::USER && n->o1->kind == Node::VAR) {
+                if (n->value_type != n->o1->value_type || n->user_type != n->o1->user_type) {
+                    throw std::string("types of objects '"
+                                      + n->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->value_type == Node::USER ? (n->user_type) : (n->value_type == Node::INTEGER ? "int"
+                                                                                                           : "float"))
+                                      + ") and '"
+                                      + n->o1->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->o1->value_type == Node::USER ? (n->o1->user_type) : (n->o1->value_type == Node::INTEGER
+                                                                                    ? "int" : "float"))
+                                      + ") does not match!"
+                    );
+                }
 
-            if (table.at(n->var_name)->getType() != val->getType()) {
-                if (table.at(n->var_name)->getType()->isIntegerTy(32) && val->getType()->isDoubleTy()) {
-                    Value *buf = val;
-                    val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                builder->CreateMemCpy(
+                        table.at(n->var_name),
+                        table.at(n->o1->var_name),
+                        module->getDataLayout().getTypeAllocSize(table.at(n->o1->var_name)->getType()),
+                        table.at(n->o1->var_name)->getPointerAlignment(module->getDataLayout())
+                );
+            } else if (n->o1->value_type == Node::USER && (n->o1->kind == Node::FUNCTION_CALL || n->o1->kind == Node::METHOD_CALL)) {
+                if (n->value_type != n->o1->value_type || n->user_type != n->o1->user_type) {
+                    throw std::string("types of objects '"
+                                      + n->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->value_type == Node::USER ? (n->user_type) : (n->value_type == Node::INTEGER ? "int"
+                                                                                                           : "float"))
+                                      + ") and '"
+                                      + n->o1->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->o1->value_type == Node::USER ? (n->o1->user_type) : (n->o1->value_type == Node::INTEGER
+                                                                                    ? "int" : "float"))
+                                      + ") does not match!"
+                    );
                 }
-                else if (table.at(n->var_name)->getType()->isDoubleTy() && val->getType()->isIntegerTy(32)) {
-                    Value *buf = val;
-                    val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+
+                generate(n->o1);
+                Value *call = stack.top();
+                stack.pop();
+
+                builder->CreateMemCpy(
+                        table.at(n->var_name),
+                        call,
+                        module->getDataLayout().getTypeAllocSize(call->getType()),
+                        call->getPointerAlignment(module->getDataLayout())
+                );
+            } else {
+                if (n->o1->kind == Node::OBJECT_CONSTRUCT) {
+                    stack.emplace(table.at(n->var_name));
                 }
+
+                generate(n->o1); // generate initial value
+
+                if (n->o1->kind == Node::OBJECT_CONSTRUCT) { // object constructors are void functions
+                    break;
+                }
+
+                Value *val = stack.top(); // take it from the stack
+                stack.pop(); // erase it from the stack
+
+                if (table.at(n->var_name)->getType() != val->getType()) {
+                    if (table.at(n->var_name)->getType()->isIntegerTy(32) && val->getType()->isDoubleTy()) {
+                        Value *buf = val;
+                        val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                    } else if (table.at(n->var_name)->getType()->isDoubleTy() && val->getType()->isIntegerTy(32)) {
+                        Value *buf = val;
+                        val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+                    }
+                }
+
+                builder->CreateStore(val, table.at(n->var_name));
             }
-            builder->CreateStore(val, table.at(n->var_name));
             break;
         }
         case Node::DELETE:
             table.erase(n->var_name); // erase pointer from the table
+            last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), n->var_name));
             break;
         case Node::VAR: { // push value of the variable to the stack
             Type *type;
-            if(table.at(n->var_name)->getType() == Type::getInt32PtrTy(context))
-                type = Type::getInt32Ty(context);
-            else if(table.at(n->var_name)->getType() == Type::getDoublePtrTy(context))
-                type = Type::getDoubleTy(context);
 
-            stack.push(builder->CreateLoad(type, table.at(n->var_name), n->var_name));
+            auto ptr = table.at(n->var_name);
+            for (auto &&user_type : user_types) {
+                auto iter = std::find(
+                        std::begin(user_type.second.second.first),
+                        std::end(user_type.second.second.first),
+                        n->var_name
+                );
+                if (iter != std::end(user_type.second.second.first)) {
+                    ptr = builder->CreateGEP(
+                            user_type.second.first,
+                            builder->CreateLoad(table.at("this")),
+                            {
+                                    ConstantInt::get(Type::getInt32Ty(context), 0),
+                                    ConstantInt::get(
+                                            Type::getInt32Ty(context),
+                                            static_cast<uint64_t>(std::distance(std::begin(user_type.second.second.first), iter))
+                                    )
+                            },
+                            n->var_name
+                    );
+                }
+            }
+
+            if (table.at(n->var_name)->getType() == Type::getInt32PtrTy(context)) {
+                type = Type::getInt32Ty(context);
+            } else if (table.at(n->var_name)->getType() == Type::getDoublePtrTy(context)) {
+                type = Type::getDoubleTy(context);
+            } else {
+                stack.emplace(table.at(n->var_name));
+                break;
+            }
+
+            stack.emplace(builder->CreateLoad(type, ptr, n->var_name));
+            break;
+        }
+        case Node::FUNC_OBJ_PROPERTY_ACCESS: {
+            Value *ptr;
+
+            auto user_type = user_types.at(n->user_type);
+            auto iter = std::find(
+                    std::begin(user_type.second.first),
+                    std::end(user_type.second.first),
+                    n->property_name
+            );
+            int property_access = user_type.second.second.at(static_cast<unsigned>(std::distance(std::begin(user_type.second.first), iter)));
+
+            if (property_access == Node::PRIVATE) {
+                throw std::string(" property '" + n->property_name + "' of object '" + n->var_name + "' is private");
+            }
+
+            generate(n->o1);
+            Value* src = stack.top();
+            stack.pop();
+
+            while (cast<PointerType>(src->getType())->getElementType()->isPointerTy()) {
+                Value *temp = builder->CreateLoad(src);
+                src = temp;
+            }
+
+            if (iter != std::end(user_type.second.first)) {
+                ptr = builder->CreateGEP(
+                        src,
+                        {
+                                ConstantInt::get(Type::getInt32Ty(context), 0),
+                                ConstantInt::get(
+                                        Type::getInt32Ty(context),
+                                        static_cast<uint64_t>(std::distance(std::begin(user_type.second.first), iter))
+                                )
+                        },
+                        n->var_name + "::" + n->property_name
+                );
+            }
+            stack.emplace(builder->CreateLoad(ptr, n->property_name));
+            break;
+        }
+        case Node::PROPERTY_ACCESS: {
+            Value *ptr;
+            auto user_type = user_types.at(n->user_type);
+            auto iter = std::find(
+                    std::begin(user_type.second.first),
+                    std::end(user_type.second.first),
+                    n->property_name
+            );
+            int property_access = user_type.second.second.at(static_cast<unsigned>(std::distance(std::begin(user_type.second.first), iter)));
+
+            if (n->var_name != "this" && property_access == Node::PRIVATE) {
+                throw std::string(" property '" + n->property_name + "' of object '" + n->var_name + "' is private");
+            }
+
+            Value* src = table.at(n->var_name);
+            while (cast<PointerType>(src->getType())->getElementType()->isPointerTy()) {
+                Value *temp = builder->CreateLoad(src);
+                src = temp;
+            }
+
+            if (iter != std::end(user_type.second.first)) {
+                ptr = builder->CreateGEP(
+                        src,
+                        {
+                                ConstantInt::get(Type::getInt32Ty(context), 0),
+                                ConstantInt::get(
+                                        Type::getInt32Ty(context),
+                                        static_cast<uint64_t>(std::distance(std::begin(user_type.second.first), iter))
+                                )
+                        },
+                        n->var_name + "::" + n->property_name
+                );
+            }
+            stack.emplace(builder->CreateLoad(ptr, n->property_name));
             break;
         }
         case Node::FUNCTION_CALL: { // generate function's call
-            Function *callee = functions.at(n->var_name); // get the function's prototype
+            Function *callee = functions.at(n->var_name).first; // get the function's prototype
 
-            if (callee->arg_size() != n->call_args.size()) // check number of arguments in prototype and in calling
+            if (callee->arg_size() != n->func_call_args.size()) { // check number of arguments in prototype and in calling
                 throw std::string(
                         "Invalid number arguments (" +
-                                std::to_string(n->call_args.size()) +
-                                ") , expected " +
-                                std::to_string(callee->arg_size())
+                        std::to_string(n->func_call_args.size()) +
+                        ") , expected " +
+                        std::to_string(callee->arg_size())
                 );
+            }
 
             std::vector<Value *> args; // generate values of arguments
-            for (auto &&arg : n->call_args) {
+            for (auto &&arg : n->func_call_args) {
                 generate(arg); // generate value
-                args.push_back(stack.top()); // take it from the stack
+                args.emplace_back(stack.top()); // take it from the stack
                 stack.pop(); // erase it from the stack
             }
 
             Value* call; // generate call
-            if (callee->getReturnType() == Type::getVoidTy(context))
+            if (callee->getReturnType() == Type::getVoidTy(context)) {
                 call = builder->CreateCall(callee, args); // LLVM forbids give name to call of void function
-            else
+            } else {
                 call = builder->CreateCall(callee, args, n->var_name+"_call");
+            }
 
-            stack.push(call); // push call to the stack
+            stack.emplace(call); // push call to the stack
+            break;
+        }
+        case Node::METHOD_CALL: { // generate function's call
+            Function *callee = functions.at(n->property_name).first; // get the function's prototype
+
+            if (n->var_name != "this" && functions.at(n->property_name).second == Node::PRIVATE) {
+                throw std::string(" method '" + n->property_name + "' of object '" + n->var_name + "' is private");
+            }
+
+            if (callee->arg_size() != n->func_call_args.size() && callee->arg_size()-n->func_call_args.size()>1) { // check number of arguments in prototype and in calling
+                throw std::string(
+                        " Invalid number arguments (" +
+                        std::to_string(n->func_call_args.size()) +
+                        "), expected " +
+                        std::to_string(callee->arg_size())
+                );
+            }
+
+            std::vector<Value *> args; // generate values of arguments
+            Value *self = table.at(n->var_name);
+            while (cast<PointerType>(self->getType())->getElementType()->isPointerTy()) {
+                Value *temp = builder->CreateLoad(self);
+                self = temp;
+            }
+            args.emplace_back(self);
+
+            for (auto &&arg : n->func_call_args) {
+                generate(arg); // generate value
+                args.emplace_back(stack.top()); // take it from the stack
+                stack.pop(); // erase it from the stack
+            }
+
+            Value* call; // generate call
+            if (callee->getReturnType() == Type::getVoidTy(context)) {
+                call = builder->CreateCall(callee, args); // LLVM forbids give name to call of void function
+            } else {
+                call = builder->CreateCall(callee, args, n->var_name+"::"+n->property_name+"_call");
+            }
+
+            stack.emplace(call); // push call to the stack
+            break;
+        }
+        case Node::FUNC_OBJ_METHOD_CALL: { // generate function's call
+            generate(n->o1);
+            Value *obj = stack.top();
+            stack.pop();
+
+            Function *callee = functions.at(n->property_name).first; // get the function's prototype
+
+            if (n->var_name != "this" && functions.at(n->property_name).second == Node::PRIVATE) {
+                throw std::string(" method '" + n->property_name + "' of object returned by function '" + n->var_name + "' is private");
+            }
+
+            if (callee->arg_size() != n->func_call_args.size() && callee->arg_size()-n->func_call_args.size()>1) { // check number of arguments in prototype and in calling
+                throw std::string(
+                        " Invalid number arguments (" +
+                        std::to_string(n->func_call_args.size()) +
+                        "), expected " +
+                        std::to_string(callee->arg_size())
+                );
+            }
+
+            std::vector<Value *> args; // generate values of arguments
+            while (cast<PointerType>(obj->getType())->getElementType()->isPointerTy()) {
+                Value *temp = builder->CreateLoad(obj);
+                obj = temp;
+            }
+            args.emplace_back(obj);
+
+            for (auto &&arg : n->func_call_args) {
+                generate(arg); // generate value
+                args.emplace_back(stack.top()); // take it from the stack
+                stack.pop(); // erase it from the stack
+            }
+
+            Value* call; // generate call
+            if (callee->getReturnType() == Type::getVoidTy(context)) {
+                call = builder->CreateCall(callee, args); // LLVM forbids give name to call of void function
+            } else {
+                call = builder->CreateCall(callee, args, n->var_name+"::"+n->property_name+"_call");
+            }
+
+            stack.emplace(call); // push call to the stack
+            break;
+        }
+        case Node::OBJECT_CONSTRUCT: { // generate function's call
+            Function *callee = functions.at(n->var_name).first; // get the function's prototype
+
+            if (callee->arg_size() != n->func_call_args.size() && callee->arg_size()-n->func_call_args.size()>1) { // check number of arguments in prototype and in calling
+                throw std::string(
+                        " Invalid number arguments (" +
+                        std::to_string(n->func_call_args.size()) +
+                        "), expected " +
+                        std::to_string(callee->arg_size())
+                );
+            }
+
+            std::vector<Value *> args; // generate values of arguments
+            args.emplace_back(stack.top());
+            stack.pop();
+            for (auto &&arg : n->func_call_args) {
+                generate(arg); // generate value
+                args.emplace_back(stack.top()); // take it from the stack
+                stack.pop(); // erase it from the stack
+            }
+
+            Value* call; // generate call
+            if (callee->getReturnType() == Type::getVoidTy(context)) {
+                call = builder->CreateCall(callee, args); // LLVM forbids give name to call of void function
+            } else {
+                call = builder->CreateCall(callee, args, n->var_name+"_call");
+            }
+
+            stack.emplace(call); // push call to the stack
             break;
         }
         case Node::ARRAY_ACCESS: { // generate access to array's element
@@ -190,7 +531,7 @@ void Generator::generate(Node *n) {
             stack.pop(); // erase it from the stack
 
             Value *arr_ptr = table.at(n->var_name); // get array's pointer
-            stack.push(
+            stack.emplace(
                     builder->CreateLoad(
                             builder->CreateGEP( // get element's pointer
                                     arr_ptr,
@@ -203,21 +544,21 @@ void Generator::generate(Node *n) {
                             n->var_name
                     )
             );
-
             break;
         }
         case Node::CONST: { // generate constant value
             switch (n->value_type) {
-                case Node::string: // str constant
-                    stack.push(builder->CreateGlobalStringPtr(n->str_val));
+                case Node::STRING: // str constant
+                    stack.emplace(builder->CreateGlobalStringPtr(n->str_val));
                     break;
-                case Node::integer: { // integer constant
-                    stack.push(ConstantInt::get(context, APInt(32, static_cast<uint64_t>(n->int_val))));
+                case Node::INTEGER: { // integer constant
+                    stack.emplace(ConstantInt::get(context, APInt(32, static_cast<uint64_t>(n->int_val))));
                     break;
                 }
-                case Node::floating: // float constant
-                    stack.push(ConstantFP::get(context, APFloat(n->float_val)));
+                case Node::FLOATING: { // float constant
+                    stack.emplace(ConstantFP::get(context, APFloat(n->float_val)));
                     break;
+                }
             }
             break;
         }
@@ -235,14 +576,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToUI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateAdd(left, right)); // push generated addition to the stack
+                stack.emplace(builder->CreateAdd(left, right)); // push generated addition to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateUIToFP(buf, Type::getDoubleTy(context));
                 }
-                stack.push(builder->CreateFAdd(left, right)); // push generated addition to the stack
+                stack.emplace(builder->CreateFAdd(left, right)); // push generated addition to the stack
             }
             break;
         }
@@ -260,14 +601,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToUI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateSub(left, right)); // push generated subtraction to the stack
+                stack.emplace(builder->CreateSub(left, right)); // push generated subtraction to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateUIToFP(buf, Type::getDoubleTy(context));
                 }
-                stack.push(builder->CreateFSub(left, right)); // push generated subtraction to the stack
+                stack.emplace(builder->CreateFSub(left, right)); // push generated subtraction to the stack
             }
             break;
         }
@@ -285,14 +626,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToUI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateMul(left, right)); // push generated multiplication to the stack
+                stack.emplace(builder->CreateMul(left, right)); // push generated multiplication to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateUIToFP(buf, Type::getDoubleTy(context));
                 }
-                stack.push(builder->CreateFMul(left, right)); // push generated multiplication to the stack
+                stack.emplace(builder->CreateFMul(left, right)); // push generated multiplication to the stack
             }
             break;
         }
@@ -310,14 +651,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateSDiv(left, right)); // push generated division to the stack
+                stack.emplace(builder->CreateSDiv(left, right)); // push generated division to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                stack.push(builder->CreateFDiv(left, right)); // push generated division to the stack
+                stack.emplace(builder->CreateFDiv(left, right)); // push generated division to the stack
             }
             break;
         }
@@ -335,14 +676,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpSLT(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpSLT(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpULT(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpULT(left, right)); // push generated test to the stack
             }
             break;
         }
@@ -360,14 +701,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpSLE(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpSLE(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpULE(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpULE(left, right)); // push generated test to the stack
             }
             break;
         }
@@ -385,14 +726,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpSGT(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpSGT(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpUGT(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpUGT(left, right)); // push generated test to the stack
             }
             break;
         }
@@ -410,14 +751,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpSGE(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpSGE(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpUGE(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpUGE(left, right)); // push generated test to the stack
             }
             break;        }
         case Node::IS_TEST: { // is FIXME floating point test
@@ -434,14 +775,14 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpEQ(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpEQ(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpUEQ(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpUEQ(left, right)); // push generated test to the stack
             }
             break;
         }
@@ -459,52 +800,322 @@ void Generator::generate(Node *n) {
                     Value *buf = right;
                     right = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
                 }
-                stack.push(builder->CreateICmpNE(left, right)); // push generated test to the stack
+                stack.emplace(builder->CreateICmpNE(left, right)); // push generated test to the stack
             }
             else if (left->getType()->isDoubleTy()) { // left operand is double
                 if (right->getType()->isIntegerTy()) { // if right operand is int, we have to convert it to double
                     Value *buf = right;
                     right = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
                 }
-                //stack.push(builder->CreateFCmpUNE(left, right)); // push generated test to the stack
+                //stack.emplace(builder->CreateFCmpUNE(left, right)); // push generated test to the stack
             }
             break;
         }
         case Node::SET: { // generate an update variable's or array element's value
-            generate(n->o1); // generate new value
-            Value *val = stack.top(); // take it from the stack
-            stack.pop(); // erase it from the stack
+            if (n->o1->value_type == Node::USER && n->o1->kind == Node::VAR) {
+                if (n->value_type != n->o1->value_type || n->user_type != n->o1->user_type) {
+                    throw std::string("types of objects '"
+                                      + n->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->value_type == Node::USER ? (n->user_type) : (n->value_type == Node::INTEGER ? "int"
+                                                                                                           : "float"))
+                                      + ") and '"
+                                      + n->o1->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->o1->value_type == Node::USER ? (n->o1->user_type) : (n->o1->value_type == Node::INTEGER
+                                                                                    ? "int" : "float"))
+                                      + ") does not match!"
+                    );
+                }
 
-            if (n->o2 != NULL) { // change value of array's element
-                generate(n->o2); // generate the number of element
-                Value *elements_count_val = stack.top(); // take it from the stack
+                builder->CreateMemCpy(
+                        table.at(n->var_name),
+                        table.at(n->o1->var_name),
+                        module->getDataLayout().getTypeAllocSize(table.at(n->o1->var_name)->getType()),
+                        table.at(n->o1->var_name)->getPointerAlignment(module->getDataLayout())
+                );
+            } else if (n->o1->value_type == Node::USER && (n->o1->kind == Node::FUNCTION_CALL || n->o1->kind == Node::METHOD_CALL)) {
+                if (n->value_type != n->o1->value_type || n->user_type != n->o1->user_type) {
+                    throw std::string("types of objects '"
+                                      + n->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->value_type == Node::USER ? (n->user_type) : (n->value_type == Node::INTEGER ? "int"
+                                                                                                           : "float"))
+                                      + ") and '"
+                                      + n->o1->var_name
+                                      + "' ("
+                                      + std::string(
+                            n->o1->value_type == Node::USER ? (n->o1->user_type) : (n->o1->value_type == Node::INTEGER
+                                                                                    ? "int" : "float"))
+                                      + ") does not match!"
+                    );
+                }
+
+                generate(n->o1);
+                Value *call = stack.top();
+                stack.pop();
+
+                builder->CreateMemCpy(
+                        table.at(n->var_name),
+                        call,
+                        module->getDataLayout().getTypeAllocSize(call->getType()),
+                        call->getPointerAlignment(module->getDataLayout())
+                );
+            } else {
+                if (n->o1->kind == Node::OBJECT_CONSTRUCT) {
+                    stack.emplace(table.at(n->var_name));
+                }
+
+                generate(n->o1); // generate new value
+
+                if (n->o1->kind == Node::OBJECT_CONSTRUCT) { // object constructors are void functions
+                    break;
+                }
+
+                Value *val = stack.top(); // take it from the stack
                 stack.pop(); // erase it from the stack
 
-                Value *arr_ptr = table.at(n->var_name); // get array's pointer
-                builder->CreateStore(
-                        val,
-                        builder->CreateGEP( // get element's pointer
-                                arr_ptr,
+                if (n->o2 != nullptr) { // change value of array's element
+                    generate(n->o2); // generate the number of element
+                    Value *elements_count_val = stack.top(); // take it from the stack
+                    stack.pop(); // erase it from the stack
+
+                    Value *arr_ptr = table.at(n->var_name); // get array's pointer
+                    Value *el_ptr = builder->CreateGEP( // get element's pointer
+                            arr_ptr,
+                            {
+                                    ConstantInt::get(Type::getInt32Ty(context), 0),
+                                    elements_count_val
+                            },
+                            n->var_name
+                    );
+
+                    if (el_ptr->getType() != val->getType()) {
+                        if (el_ptr->getType() == Type::getInt32PtrTy(context) && val->getType()->isDoubleTy()) {
+                            Value *buf = val;
+                            val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                        } else if (el_ptr->getType() == Type::getDoublePtrTy(context) &&
+                                   val->getType()->isIntegerTy(32)) {
+                            Value *buf = val;
+                            val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+                        }
+                    }
+                    builder->CreateStore(val, el_ptr); // update value of variable
+
+                } else if (!n->property_name.empty()) { // change value of object's property
+                    Value *property_ptr;
+                    Value *temp = table.at(n->var_name);
+                    while (temp->getType()->isPointerTy()) {
+                        Value *_temp = builder->CreateLoad(temp);
+                        temp = _temp;
+                    }
+
+                    auto user_type = user_types.at(temp->getType()->getStructName().str());
+                    auto iter = std::find(
+                            std::begin(user_type.second.first),
+                            std::end(user_type.second.first),
+                            n->property_name
+                    );
+                    int property_access = user_type.second.second.at(static_cast<unsigned>(std::distance(std::begin(user_type.second.first),iter)));
+
+                    if (n->var_name != "this" && property_access == Node::PRIVATE) {
+                        throw std::string(" property '" + n->property_name + "' of object '" + n->var_name + "' is private");
+                    }
+
+                    Value* src = table.at(n->var_name);
+                    while (cast<PointerType>(src->getType())->getElementType()->isPointerTy()) {
+                        Value *_temp = builder->CreateLoad(src);
+                        src = _temp;
+                    }
+
+                    if (iter != std::end(user_type.second.first)) {
+                        property_ptr = builder->CreateGEP(
+                                src,
                                 {
                                         ConstantInt::get(Type::getInt32Ty(context), 0),
-                                        elements_count_val
+                                        ConstantInt::get(
+                                                Type::getInt32Ty(context),
+                                                static_cast<uint64_t>(std::distance(std::begin(user_type.second.first), iter))
+                                        )
                                 },
-                                n->var_name
-                        )
-                ); // update the value
-            } else {
-                if (table.at(n->var_name)->getType() != val->getType()) {
-                    if (table.at(n->var_name)->getType()->isIntegerTy(32) && val->getType()->isDoubleTy()) {
-                        Value *buf = val;
-                        val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                                n->var_name + "::" + n->property_name
+                        );
                     }
-                    else if (table.at(n->var_name)->getType()->isDoubleTy() && val->getType()->isIntegerTy(32)) {
-                        Value *buf = val;
-                        val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+
+                    /*if (el_ptr->getType() != val->getType()) {
+                        if (el_ptr->getType() == Type::getInt32PtrTy(context) && val->getType()->isDoubleTy()) {
+                            Value *buf = val;
+                            val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                        } else if (el_ptr->getType() == Type::getDoublePtrTy(context) &&
+                                   val->getType()->isIntegerTy(32)) {
+                            Value *buf = val;
+                            val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+                        }
+                    }*/
+                    builder->CreateStore(val, property_ptr); // update value of variable
+                } else {
+                    if (table.at(n->var_name)->getType() != val->getType()) {
+                        if (table.at(n->var_name)->getType() == Type::getInt32PtrTy(context) &&
+                            val->getType()->isDoubleTy()) {
+                            Value *buf = val;
+                            val = builder->CreateFPToSI(buf, Type::getInt32Ty(context));
+                        } else if (table.at(n->var_name)->getType() == Type::getDoublePtrTy(context) &&
+                                   val->getType()->isIntegerTy(32)) {
+                            Value *buf = val;
+                            val = builder->CreateSIToFP(buf, Type::getDoubleTy(context));
+                        }
+                    }
+
+                    auto ptr = table.at(n->var_name);
+
+                    for (auto &&user_type : user_types) {
+                        auto iter = std::find(
+                                std::begin(user_type.second.second.first),
+                                std::end(user_type.second.second.first),
+                                n->var_name
+                        );
+                        if (iter != std::end(user_type.second.second.first)) {
+                            ptr = builder->CreateGEP(
+                                    user_type.second.first,
+                                    builder->CreateLoad(table.at("this")),
+                                    {
+                                            ConstantInt::get(Type::getInt32Ty(context), 0),
+                                            ConstantInt::get(
+                                                    Type::getInt32Ty(context),
+                                                    static_cast<uint64_t>(std::distance(std::begin(user_type.second.second.first), iter))
+                                            )
+                                    },
+                                    n->var_name
+                            );
+                        }
+                    }
+                    builder->CreateStore(val, ptr); // update value of variable
+                }
+            }
+
+            break;
+        }
+        case Node::CLASS_DEFINE: {
+            std::vector<Type *> properties_types;
+            std::vector<std::string> properties_names;
+            std::vector<int> properties_access;
+            for (auto &&defProperty : n->class_def_properties) { // generate properties first
+                if (defProperty.second.second->kind == Node::NEW) {
+                    properties_names.emplace_back(defProperty.first);
+                    properties_access.emplace_back(defProperty.second.first);
+                    switch (defProperty.second.second->value_type) {
+                        case Node::INTEGER:
+                            properties_types.emplace_back(Type::getInt32Ty(context));
+                            break;
+                        case Node::FLOATING:
+                            properties_types.emplace_back(Type::getDoubleTy(context));
+                            break;
+                    }
+                    table.emplace(defProperty.first, builder->CreateAlloca(properties_types.back(), nullptr));
+                }
+            }
+
+            StructType* class_type = StructType::create(context, n->var_name);
+            class_type->setBody(properties_types);
+            user_types.emplace(n->var_name, std::make_pair(class_type, std::make_pair(properties_names, properties_access)));
+
+            for (auto &&defProperty : n->class_def_properties) { // then, generate methods
+                if (defProperty.second.second->kind == Node::FUNCTION_DEFINE) {
+                    FunctionType *type;
+
+                    // generate arguments
+                    std::vector<Type *> args_types;
+                    std::vector<std::string> args_names;
+
+                    args_types.emplace_back(PointerType::get(class_type, 0));
+                    args_names.emplace_back("this");
+
+                    for (auto &iterator : defProperty.second.second->o1->func_def_args) {
+                        switch (iterator.second->value_type) {
+                            case Node::INTEGER:
+                                args_types.emplace_back(Type::getInt32Ty(context));
+                                break;
+                            case Node::FLOATING:
+                                args_types.emplace_back(Type::getDoubleTy(context));
+                                break;
+                            case Node::USER:
+                                args_types.emplace_back(PointerType::get(user_types.at(iterator.second->user_type_name).first, 0));
+                                break;
+                        }
+                        args_names.emplace_back(iterator.first);
+                    }
+
+                    switch (defProperty.second.second->value_type) { // set type of the function
+                        case Node::INTEGER:
+                            type = FunctionType::get(Type::getInt32Ty(context), args_types, false);
+                            break;
+                        case Node::FLOATING:
+                            type = FunctionType::get(Type::getDoubleTy(context), args_types, false);
+                            break;
+                        case Node::USER:
+                            type = FunctionType::get(PointerType::get(class_type, 0), args_types, false);
+                            break;
+                        default:
+                            type = FunctionType::get(Type::getVoidTy(context), args_types, false);
+                    }
+                    Function *func = Function::Create(type, Function::ExternalLinkage, defProperty.second.second->var_name, module.get());
+                    functions.emplace(defProperty.second.second->var_name, std::make_pair(func, defProperty.second.first));
+
+                    BasicBlock *entry = BasicBlock::Create(context, "entry", func);
+                    builder->SetInsertPoint(entry); // set new insert block
+
+                    unsigned long idx = 0;
+                    for (auto &Arg : func->args()) { // create pointers to arguments of the function
+                        std::string name = args_names.at(idx++);
+                        Arg.setName(name);
+
+                        // insert argument's allocator to the table
+                        table.emplace(
+                                name,
+                                builder->CreateAlloca(
+                                        Arg.getType(),
+                                        nullptr,
+                                        name+"_ptr"
+                                )
+                        );
+                        builder->CreateStore(&Arg, table.at(name)); // store the value of argument to allocator
+                    }
+
+                    generate(defProperty.second.second->o2); // generate body of the function
+
+                    if (!func->getAttributes().hasAttribute(0, "ret")) {
+                        switch (defProperty.second.second->value_type) { // create default return value
+                            case Node::INTEGER:
+                                builder->CreateRet(ConstantInt::get(Type::getInt32Ty(context), 0));
+                                break;
+                            case Node::FLOATING:
+                                builder->CreateRet(ConstantFP::get(Type::getDoubleTy(context), 0.0));
+                                break;
+                            default:
+                                builder->CreateRetVoid();
+                        }
+                    }
+
+                    passmgr->run(*func); // run the optimizer
+
+                    for (auto &&var : last_vars) { // erase vars declared in this function
+                        table.erase(var);
+                        last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+                    }
+
+                    for (auto &Arg : func->args()) { // erase arguments' allocators from the table
+                        table.erase(Arg.getName());
                     }
                 }
-                builder->CreateStore(val, table.at(n->var_name)); // update value of variable
             }
+
+            for (auto &&item : properties_names) {
+                table.erase(item);
+            }
+
             break;
         }
         case Node::IF: { // generate 'if' condition without 'else' branch
@@ -519,7 +1130,18 @@ void Generator::generate(Node *n) {
             builder->CreateCondBr(stack.top(), thenBlock, elseBlock); // create conditional goto
 
             builder->SetInsertPoint(thenBlock);
+
+            std::vector<std::string> _temp(last_vars);
+            last_vars.clear();
+
             generate(n->o2); // generate the body of 'then' branch
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = _temp;
 
             builder->CreateBr(mergeBlock);
 
@@ -545,13 +1167,35 @@ void Generator::generate(Node *n) {
             builder->CreateCondBr(stack.top(), thenBlock, elseBlock); // create conditional goto
 
             builder->SetInsertPoint(thenBlock);
+
+            std::vector<std::string> _temp(last_vars);
+            last_vars.clear();
+
             generate(n->o2); // generate the body of 'then' branch
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = _temp;
 
             builder->CreateBr(mergeBlock);
 
             parent->getBasicBlockList().push_back(elseBlock);
             builder->SetInsertPoint(elseBlock);
+
+            std::vector<std::string> __temp(last_vars);
+            last_vars.clear();
+
             generate(n->o3); // generate the body of 'else' branch
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = __temp;
 
             builder->CreateBr(mergeBlock);
 
@@ -566,11 +1210,21 @@ void Generator::generate(Node *n) {
             builder->CreateBr(loopBlock);
 
             builder->SetInsertPoint(loopBlock);
-            generate(n->o1); // generate the body
 
+            std::vector<std::string> _temp(last_vars);
+            last_vars.clear();
+
+            generate(n->o1); // generate the body
             generate(n->o2); // generate the condition
             Value *condition = stack.top(); // take it from the stack
             stack.pop(); // erase it from the stack
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = _temp;
 
             BasicBlock *afterBlock = BasicBlock::Create(context, "afterloop", parent);
             builder->CreateCondBr(condition, loopBlock, afterBlock); // create conditional goto
@@ -584,7 +1238,18 @@ void Generator::generate(Node *n) {
             builder->CreateBr(loopBlock);
 
             builder->SetInsertPoint(loopBlock);
+
+            std::vector<std::string> _temp(last_vars);
+            last_vars.clear();
+
             generate(n->o2); // generate the body
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = _temp;
 
             generate(n->o1); // generate the condition
             Value *condition = stack.top(); // take it from the stack
@@ -599,9 +1264,7 @@ void Generator::generate(Node *n) {
         case Node::REPEAT: { // 'repeat' cycle FIXME
             if (table.count("index") == 0)
             {
-                table.insert(std::pair<std::string, Value *>("index", builder->CreateAlloca(Type::getInt32Ty(context),
-                                                                                            nullptr,
-                                                                                            "index_ptr")));
+                table.emplace("index", builder->CreateAlloca(Type::getInt32Ty(context), nullptr, "index_ptr"));
             }
 
             builder->CreateStore(ConstantInt::get(Type::getInt32Ty(context), APInt(32, 0)),
@@ -611,7 +1274,18 @@ void Generator::generate(Node *n) {
             BasicBlock *loopBlock = BasicBlock::Create(context, "loop", parent);
             builder->CreateBr(loopBlock); // go to begin of the loop
             builder->SetInsertPoint(loopBlock);
-            generate(n->o2); // generate body of the loop
+
+            std::vector<std::string> _temp(last_vars);
+            last_vars.clear();
+
+            generate(n->o2); // generate the body
+
+            for (auto &&var : last_vars) {
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            last_vars = _temp;
 
             generate(n->o1); // generate the condition
             Value *times = stack.top(); // take it form the stack
@@ -641,34 +1315,41 @@ void Generator::generate(Node *n) {
             // generate arguments
             std::vector<Type *> args_types;
             std::vector<std::string> args_names;
-            for (auto &iterator : n->o1->def_args) {                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
-                switch (iterator.second) {
-                    case Node::integer:
-                        args_types.push_back(Type::getInt32Ty(context));
+            for (auto &iterator : n->o1->func_def_args) {
+                switch (iterator.second->value_type) {
+                    case Node::INTEGER:
+                        args_types.emplace_back(Type::getInt32Ty(context));
                         break;
-                    case Node::floating:
-                        args_types.push_back(Type::getDoubleTy(context));
+                    case Node::FLOATING:
+                        args_types.emplace_back(Type::getDoubleTy(context));
+                        break;
+                    case Node::USER:
+                        args_types.emplace_back(PointerType::get(
+                                user_types.at(iterator.second->user_type_name).first, 0)
+                        );
                         break;
                 }
-                args_names.push_back(iterator.first);
+                args_names.emplace_back(iterator.first);
             }
 
             switch (n->value_type) { // set type of the function
-                case Node::integer:
+                case Node::INTEGER:
                     type = FunctionType::get(Type::getInt32Ty(context), args_types, false);
                     break;
-                case Node::floating:
+                case Node::FLOATING:
                     type = FunctionType::get(Type::getDoubleTy(context), args_types, false);
+                    break;
+                case Node::USER:
+                    type = FunctionType::get(PointerType::get(user_types.at(n->user_type).first, 0), args_types, false);
                     break;
                 default:
                     type = FunctionType::get(Type::getVoidTy(context), args_types, false);
             }
-            Function *func = Function::Create(type, Function::ExternalLinkage, n->var_name, module);
-            functions.insert(std::pair<std::string, Function*>(n->var_name, func));
+            Function *func = Function::Create(type, Function::ExternalLinkage, n->var_name, module.get());
+            functions.emplace(n->var_name, std::make_pair(func, Node::PUBLIC));
 
             BasicBlock *entry = BasicBlock::Create(context, "entry", func);
             builder->SetInsertPoint(entry); // set new insert block
-            currentBlock = entry;
 
             unsigned long idx = 0;
             for (auto &Arg : func->args()) { // create pointers to arguments of the function
@@ -676,10 +1357,13 @@ void Generator::generate(Node *n) {
                 Arg.setName(name);
 
                 // insert argument's allocator to the table
-                table.insert(std::pair<std::string, Value *>(name, builder->CreateAlloca(
-                        Arg.getType(),
-                        nullptr,
-                        name+"_ptr"))
+                table.emplace(
+                        name,
+                        builder->CreateAlloca(
+                                Arg.getType(),
+                                nullptr,
+                                name+"_ptr"
+                        )
                 );
                 builder->CreateStore(&Arg, table.at(name)); // store the value of argument to allocator
             }
@@ -688,10 +1372,10 @@ void Generator::generate(Node *n) {
 
             if (!func->getAttributes().hasAttribute(0, "ret")) {
                 switch (n->value_type) { // create default return value
-                    case Node::integer:
+                    case Node::INTEGER:
                         builder->CreateRet(ConstantInt::get(Type::getInt32Ty(context), 0));
                         break;
-                    case Node::floating:
+                    case Node::FLOATING:
                         builder->CreateRet(ConstantFP::get(Type::getDoubleTy(context), 0.0));
                         break;
                     default:
@@ -699,8 +1383,16 @@ void Generator::generate(Node *n) {
                 }
             }
 
-            for (auto &Arg : func->args()) // erase arguments' allocators from the table
+            passmgr->run(*func); // run the optimizer
+
+            for (auto &&var : last_vars) { // erase vars declared in this function
+                table.erase(var);
+                last_vars.erase(std::find(std::cbegin(last_vars), std::cend(last_vars), var));
+            }
+
+            for (auto &Arg : func->args()) { // erase arguments' allocators from the table
                 table.erase(Arg.getName());
+            }
 
             break;
         }
@@ -721,36 +1413,39 @@ void Generator::generate(Node *n) {
 
             break;
         case Node::PRINT: { // print something TODO: migrate to call of prinf function
-            if (!io_using)
+            if (!io_using) {
                 use_io();
+            }
 
             generate(n->o1); // generate the value to print
             std::vector<Value*> args;
 
             Value *format;
-            if (stack.top()->getType()->isIntegerTy(32)) // print integer
+            if (stack.top()->getType()->isIntegerTy(32)) { // print integer
                 format = table.at("int_out_format");
-            else if (stack.top()->getType()->isDoubleTy()) // print float
+            } else if (stack.top()->getType()->isDoubleTy()) { // print float
                 format = table.at("float_out_format");
-            else // print string
+            } else { // print string
                 format = table.at("str_out_format");
+            }
 
-            args.push_back(format);
-            args.push_back(stack.top()); // get the value from the stack
+            args.emplace_back(format);
+            args.emplace_back(stack.top()); // get the value from the stack
             stack.pop(); // erase it from the stack
             builder->CreateCall(printf, args); // call prinf
 
             break;
         }
         case Node::INPUT: {
-            if (!io_using)
+            if (!io_using) {
                 use_io();
+            }
 
             std::vector<Value*> args;
             Value *format = table.at("float_in_format");
 
-            args.push_back(format);
-            args.push_back(table.at(n->var_name));
+            args.emplace_back(format);
+            args.emplace_back(table.at(n->var_name));
             builder->CreateCall(scanf, args); // call scanf
 
             break;
